fix: Restrict disk and folder removal in Remove-Lab to HyperV virtualization engine

+++ AutomatedLabCore/functions/Core/Remove-Lab.ps1
@@ -149,7 +149,7 @@
                         }
                     }
 
-                    if ($Script:data.Target.Path)
+                    if ($Script:data.Target.Path -and (Get-Lab).DefaultVirtualizationEngine -eq 'HyperV')
                     {
                         $diskPath = (Join-Path -Path $Script:data.Target.Path -ChildPath Disks)
                         #Only remove disks folder if empty
@@ -161,7 +161,7 @@
                 }
 
                 #Only remove folder for VMs if folder is empty
-                if ($Script:data.Target.Path -and (Test-Path -Path $Script:data.Target.Path) -and (-not (Get-ChildItem -Path $Script:data.Target.Path)))
+                if ($Script:data.Target.Path -and (Get-Lab).DefaultVirtualizationEngine -eq 'HyperV' -and (Test-Path -Path $Script:data.Target.Path) -and (-not (Get-ChildItem -Path $Script:data.Target.Path)))
                 {
                     Remove-Item -Path $Script:data.Target.Path -Recurse -Force -Confirm:$false
                 }
